refactor: change fiel structure
