+++ src/server/endpoints/covidcast_meta.py
@@ -5,7 +5,7 @@ from typing import Optional, List, Dict
 
 from .._common import app, db
 from .._validate import extract_strings, filter_fields
-from .._printer import send_rows
+from .._printer import create_printer
 
 bp = Blueprint("covidcast_meta", __name__)
 
@@ -76,4 +76,4 @@ def handle():
     signals = [SourceSignal(v) for v in (extract_strings("signals") or [])]
     geo_types = extract_strings("geo_types")
 
-    return send_rows(filter_fields(fetch_data(time_types, geo_types, signals)))
+    return create_printer()(filter_fields(fetch_data(time_types, geo_types, signals)))
